ensure that we have. deck object
Now that gliders are pre-configured even when there is no tracking we have a situation where they do not start updating until the page is reloaded. This is caused by the deck structure (binary track) not being initialised on the client which means the newly received points have nowhere to go. Send empty one from server and create on receive.